xamp/requirements.txt
Django==5.2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
PyJWT==2.8.0
drf-yasg>=1.21.7
pyotp==2.9.0
social-auth-app-django==5.4.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0